Linear adaptation working, but with considerable overheads

+++ opta/ic.h
@@ -10,7 +10,7 @@
 void atom_func_start(uint8_t func_id);
 void atom_func_end(uint8_t func_id);
 void atom_func_start_linear(uint8_t func_id, uint16_t param);
-void atom_func_end_linear(uint8_t func_id, uint8_t param);
+void atom_func_end_linear(uint8_t func_id, uint16_t param);
 
 // Configs
 // Atomic function IDs
